Fixing file open fstream option

+++ src/adcirc/geometry/boundary.cpp
@@ -359,7 +359,7 @@ vector<string> Boundary::toStringList() {
     outputList.push_back(boost::str(boost::format("%11i %11i") %
                                     this->length() % this->boundaryCode()));
 
-  for (int i = 0; i < this->length(); i++) {
+  for (size_t i = 0; i < this->length(); i++) {
     if (this->boundaryCode() == 0 || this->boundaryCode() == 1 ||
         this->boundaryCode() == 2 || this->boundaryCode() == 10 ||
         this->boundaryCode() == 11 || this->boundaryCode() == 12 ||
